refactor: Global Provider들의 사용 위치를 index > App으로 변경
